+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.0 - 2025-07-26
+
+### Added
+
+- Added no_std support by disabling the new `std` feature that is enabled by default. Custom traits and default
+  implementation for &[u8] and &mut [u8] are provided.
+
 ## 0.5.1 - 2025-07-25
 
 ### Fixed